+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 - PEP 484 support (most of the important features of it). Thanks Claude! (@reinhrst)
 - Added ``get_line_code`` to ``Definition`` and ``Completion`` objects.
 - Again a lot of internal changes.
+- Added support for Google docstrings.
 
 0.9.0 (2015-04-10)
 ++++++++++++++++++